Fixed issue with default.ovpn stopping container from running
container was being stopped due to sed not liking default.ovpn symlink. Updated so that sed skipped this symlink. container was then also being stopped due to updateConfigs.sh script erroring if default.ovpn existed when trying to create symlink. Added option 'f' so that the symlink would be recreated even if existed