Make sure configure-openvpn.sh cannot change the current directory
VYPRVPN sets $PWD to a deleted directory, which causes a cascade of
issues for the rest of `start.sh`.

This saves the current directory before sourcing `configure-openvpn.sh`
scripts, making sure such an issue cannot happen for any VPN.

A proper fix would be to avoid sourcing the script in the first place.