fix: force PEER_DNS servers through the vpn tunnel (#1819)
* docs: add PEER_DNS documentation

#1713 Added dnsleak protections, but didn't document the escape hatch to
disable it.  This adds that documentation.

* fix: pin routes for DNS if PEER_DNS is active.

See #1667 for context and user's impacted; PIA for example advertises DNS
servers in the 10.0.0.0/8 range (10.0.0.243 for example).  If the home
LAN is 10.0.0.0/8- common- they'll have `LOCAL_NETWORK=10.0.0.0/8` which
will pin a prefer to route through the eth rather than tunnel.  The end
result being that DNS is unreachable since the tunnel must be used to reach
said DNS.

This commit modifies tunnelUp.sh such if we honor DNS routes given by the
provider, we force the route to that IP to go through the tunnel; this
overrides any local net range conflict that may occur.

Finally, this behaviour can be suppressed via `PEER_DNS_PIN_ROUTES=no`.

* fix: Add explicit dockerfile env defaults for PEER_DNS*