Fix sed regex for container reboots
The docker container can get a new IP address when it restarts. When this happens then the proxy cannot bind to the correct listening address. Replacing the 127.* IP only works the first time as the config is written on the anonymous volume. Before this change the container needs to be stopped and the anonymous volume needs to be removed with docker rm -v to reset the config to listen address 127.0.0.1.