Prevent error "Too many open files (24)"
When having many torrents downloading at the same time, Transmission is known
to throw the error "Too many open files (24)", and stop torrents. This happens
when it hits the process limit for number of file descriptors simulatenously open.

This is because, in this Docker container, the launch of transmission as user `abc`
entails a default limit of 1024 open files.

Create the file `/etc/security/limits.d/transmission.conf` and adding some lines
for that user increases this limit to 65536 soft and 1048576 hard, which should
be enough for most use cases.